Move logic out of main
