+++ src/RogueCollectionSdl/RogueCollection.vcxproj
@@ -96,7 +96,7 @@
       </PrecompiledHeader>
       <WarningLevel>Level3</WarningLevel>
       <Optimization>Disabled</Optimization>
-      <PreprocessorDefinitions>WIN32;_DEBUG;_WINDOWS;_CRT_SECURE_NO_WARNINGS;ROGOMATIC_PROTOCOL_DEBUGGING;%(PreprocessorDefinitions)</PreprocessorDefinitions>
+      <PreprocessorDefinitions>WIN32;_DEBUG;_WINDOWS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <SDLCheck>true</SDLCheck>
       <AdditionalIncludeDirectories>$(SolutionDir)lib\sdl\include\;$(SolutionDir)lib\SDL2_ttf-2.0.14\include\;$(SolutionDir)lib\SDL2_image-2.0.1\include\;$(SolutionDir)lib\NativeFileDialog\src\include\;$(SolutionDir)src\MyCurses\;$(SolutionDir)src\Shared\;$(SolutionDir)src\Shared\Frontend\;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <TreatWarningAsError>true</TreatWarningAsError>
@@ -158,7 +158,7 @@ xcopy "$(SolutionDir)lib\SDL2_ttf-2.0.14\lib\$(PlatformTarget)\*.dll" "$(OutDir)
       <Optimization>MaxSpeed</Optimization>
       <FunctionLevelLinking>true</FunctionLevelLinking>
       <IntrinsicFunctions>true</IntrinsicFunctions>
-      <PreprocessorDefinitions>WIN32;NDEBUG;_WINDOWS;_CRT_SECURE_NO_WARNINGS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
+      <PreprocessorDefinitions>WIN32;NDEBUG;_WINDOWS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <SDLCheck>true</SDLCheck>
       <AdditionalIncludeDirectories>$(SolutionDir)lib\sdl\include\;$(SolutionDir)lib\SDL2_ttf-2.0.14\include\;$(SolutionDir)lib\SDL2_image-2.0.1\include\;$(SolutionDir)lib\NativeFileDialog\src\include\;$(SolutionDir)src\MyCurses\;$(SolutionDir)src\Shared\;$(SolutionDir)src\Shared\Frontend\;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <TreatWarningAsError>true</TreatWarningAsError>
