+++ src/Shared/Frontend/start_process.h
@@ -1,0 +1,4 @@
+#pragma once
+struct Args;
+
+int StartProcess(int(*start)(Args& args), int argc, char** argv);
